+++ praticas/gestao-financeira/constants/colors.js
@@ -1,5 +1,5 @@
 export const colors = {
-  primary: "#37BF81",
+  primary: "#376dbf",
   primaryContrast: "#FFFFFF",
   inactive: "#B1B1B1",
   background: "#F5F5F5",
@@ -10,6 +10,8 @@ export const colors = {
   categoryTravel: "#82C9DE",
   primaryText: "#666666",
   secondaryText: "#B1B1B1",
-  positiveText: "#37BF81",
+  positiveText: "#376dbf",
   negativesText: "#DA5567"
 }
+
+// "#37BF81", verde feio
